Add specs for AfterFinishHook

+++ app/models/after_finish_hook.rb
@@ -18,6 +18,8 @@ class AfterFinishHook
   end
 
   def upload_log_file
+    return unless File.exist?(log_path)
+
     run.log_file.attach(
       io: File.open(log_path),
       filename: File.basename(log_path),
